requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
et-xmlfile==1.1.0
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
mypy==0.982
mypy-extensions==0.4.3
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
yarl==1.8.1
